backend/requirements.txt
# Core API Framework
fastapi
uvicorn
pydantic

# GitHub API Integration
PyGithub
requests

# ML & Data Science
scikit-learn>=1.3.0
pandas>=2.0.0
numpy
joblib
matplotlib
seaborn

# Feature Engineering
python-dateutil

# Environment & Config
python-dotenv
pydantic-settings

# CORS for frontend integration
python-multipart